+++ apps/meteor/client/lib/VideoConfManager.ts
@@ -1,7 +1,5 @@
 import type { CallPreferences, DirectCallData, DirectCallParams, IRoom, IUser, ProviderCapabilities } from '@rocket.chat/core-typings';
 import { Emitter } from '@rocket.chat/emitter';
-import { Meteor } from 'meteor/meteor';
-import { Tracker } from 'meteor/tracker';
 
 import { getConfig } from './utils/getConfig';
 import { sdk } from '../../app/utils/client/lib/SDKClient';
@@ -302,19 +300,18 @@ export const VideoConfManager = new (class VideoConfManager extends Emitter<Vide
 		return false;
 	}
 
-	public updateUser(): void {
-		const userId = Meteor.userId();
-
+	public updateUser(userId: string | null, isLoggingIn: boolean, isConnected: boolean): void {
 		this.debugLog(`[VideoConf] Logged user or connection status has changed.`);
 
 		if (this.userId) {
 			this.disconnect(this.userId !== userId);
 		}
 
-		if (!Meteor.status().connected || (userId && Meteor.loggingIn())) {
+		if (!isConnected || (userId && isLoggingIn)) {
 			this.debugLog(`[VideoConf] Connection lost or login process still pending, skipping user change.`);
 			return;
 		}
+
 		if (userId) {
 			this.connectUser(userId);
 		}
@@ -713,6 +710,11 @@ export const VideoConfManager = new (class VideoConfManager extends Emitter<Vide
 			return;
 		}
 
+		if (params.callId !== this.currentCallData?.callId) {
+			this.debugLog(`[VideoConf] User ${params.uid} has joined the call ${params.callId}, but we aren't calling with that id.`);
+			return;
+		}
+
 		this.infoLog(`[VideoConf] User ${params.uid} has joined a call we started ${params.callId}.`);
 		this.onDirectCallAccepted(params, true);
 	}
@@ -783,5 +785,3 @@ export const VideoConfManager = new (class VideoConfManager extends Emitter<Vide
 		return this.dismissedCalls.has(callId);
 	}
 })();
-
-Meteor.startup(() => Tracker.autorun(() => VideoConfManager.updateUser()));