+++ packages/apps-engine/src/definition/metadata/AppMethod.ts
@@ -83,6 +83,7 @@ export enum AppMethod {
      * @deprecated please use the AppMethod.EXECUTE_POST_LIVECHAT_ROOM_CLOSED method
      */
     EXECUTE_LIVECHAT_ROOM_CLOSED_HANDLER = 'executeLivechatRoomClosedHandler',
+    EXECUTE_PRE_LIVECHAT_ROOM_CREATE_PREVENT = 'executeLivechatRoomCreatePrevent',
     EXECUTE_POST_LIVECHAT_ROOM_CLOSED = 'executePostLivechatRoomClosed',
     EXECUTE_POST_LIVECHAT_AGENT_ASSIGNED = 'executePostLivechatAgentAssigned',
     EXECUTE_POST_LIVECHAT_AGENT_UNASSIGNED = 'executePostLivechatAgentUnassigned',