+++ packages/model-typings/src/models/IUsersModel.ts
@@ -19,6 +19,7 @@ import type {
 	DeleteResult,
 	WithId,
 	UpdateOptions,
+	ClientSession,
 } from 'mongodb';
 
 import type { FindPaginated, IBaseModel } from './IBaseModel';
@@ -136,7 +137,7 @@ export interface IUsersModel extends IBaseModel<IUser> {
 
 	getUserLanguages(): any;
 
-	updateStatusText(_id: any, statusText: any): any;
+	updateStatusText(_id: any, statusText: any, options?: { session?: ClientSession }): any;
 
 	updateStatusByAppId(appId: any, status: any): any;
 
@@ -216,7 +217,9 @@ export interface IUsersModel extends IBaseModel<IUser> {
 		}: { statusDefault?: string; status: UserStatus; statusConnection: UserStatus; statusText?: string },
 	): Promise<UpdateResult>;
 
-	setFederationAvatarUrlById(userId: string, federationAvatarUrl: string): Promise<void>;
+	updateStatusAndStatusDefault(userId: string, status: UserStatus, statusDefault: UserStatus): Promise<UpdateResult>;
+
+	setFederationAvatarUrlById(userId: IUser['_id'], federationAvatarUrl: string): Promise<UpdateResult>;
 
 	findSearchedServerNamesByUserId(userId: string): Promise<string[]>;
 
@@ -358,13 +361,13 @@ export interface IUsersModel extends IBaseModel<IUser> {
 	updateLastLoginById(userId: string): Promise<UpdateResult>;
 	addPasswordToHistory(userId: string, password: string, passwordHistoryAmount: number): Promise<UpdateResult>;
 	setServiceId(userId: string, serviceName: string, serviceId: string): Promise<UpdateResult>;
-	setUsername(userId: string, username: string): Promise<UpdateResult>;
-	setEmail(userId: string, email: string, verified?: boolean): Promise<UpdateResult>;
+	setUsername(userId: string, username: string, options?: { session?: ClientSession }): Promise<UpdateResult>;
+	setEmail(userId: string, email: string, verified?: boolean, options?: { session?: ClientSession }): Promise<UpdateResult>;
 	setEmailVerified(userId: string, email: string): Promise<UpdateResult>;
-	setName(userId: string, name: string): Promise<UpdateResult>;
-	unsetName(userId: string): Promise<UpdateResult>;
+	setName(userId: string, name: string, options?: { session?: ClientSession }): Promise<UpdateResult>;
+	unsetName(userId: string, options?: { session?: ClientSession }): Promise<UpdateResult>;
 	setCustomFields(userId: string, customFields: Record<string, unknown>): Promise<UpdateResult>;
-	setAvatarData(userId: string, origin: string, etag?: Date | null | string): Promise<UpdateResult>;
+	setAvatarData(userId: string, origin: string, etag?: Date | null | string, options?: { session?: ClientSession }): Promise<UpdateResult>;
 	unsetAvatarData(userId: string): Promise<UpdateResult>;
 	setUserActive(userId: string, active: boolean): Promise<UpdateResult>;
 	setAllUsersActive(active: boolean): Promise<UpdateResult | Document>;