+++ packages/ui-voip/src/providers/VoipProvider.tsx
@@ -8,7 +8,7 @@ import {
 	useToastMessageDispatch,
 } from '@rocket.chat/ui-contexts';
 import type { ReactNode } from 'react';
-import { useEffect, useMemo, useRef } from 'react';
+import { useCallback, useEffect, useMemo } from 'react';
 import { createPortal } from 'react-dom';
 import { useTranslation } from 'react-i18next';
 
@@ -39,7 +39,12 @@ const VoipProvider = ({ children }: { children: ReactNode }) => {
 	const dispatchToastMessage = useToastMessageDispatch();
 
 	// Refs
-	const remoteAudioMediaRef = useRef<HTMLAudioElement>(null);
+	const remoteAudioMediaRef = useCallback(
+		(node: HTMLMediaElement | null) => {
+			voipClient?.switchAudioElement(node);
+		},
+		[voipClient],
+	);
 
 	useEffect(() => {
 		if (!voipClient) {
@@ -54,10 +59,6 @@ const VoipProvider = ({ children }: { children: ReactNode }) => {
 		const onCallEstablished = async (): Promise<void> => {
 			voipSounds.stopAll();
 			window.addEventListener('beforeunload', onBeforeUnload);
-
-			if (voipClient.isCallee() && remoteAudioMediaRef.current) {
-				voipClient.switchMediaRenderer({ remoteMediaElement: remoteAudioMediaRef.current });
-			}
 		};
 
 		const onNetworkDisconnected = (): void => {
@@ -120,11 +121,13 @@ const VoipProvider = ({ children }: { children: ReactNode }) => {
 	}, [dispatchToastMessage, setStorageRegistered, t, voipClient, voipSounds]);
 
 	const changeAudioOutputDevice = useEffectEvent(async (selectedAudioDevice: Device): Promise<void> => {
-		if (!remoteAudioMediaRef.current) {
+		const element = voipClient?.getAudioElement();
+		if (!element) {
+			console.warn(`Failed to change audio output device: missing audio element reference.`);
 			return;
 		}
 
-		setOutputMediaDevice({ outputDevice: selectedAudioDevice, HTMLAudioElement: remoteAudioMediaRef.current });
+		setOutputMediaDevice({ outputDevice: selectedAudioDevice, HTMLAudioElement: element });
 	});
 
 	const changeAudioInputDevice = useEffectEvent(async (selectedAudioDevice: Device): Promise<void> => {